Change RDF Turtle parser so that it passes each generated triple to a user-supplied handler procedure, rather than accumulating a monolithic list of triples.  Update its tests accordingly.
Currently, it uses parameterized parsers to record the `current'
subject and predicate of the triples being generated.  I am not sure
whether this is a good idea; it leads to the construction of many, many
parsers during parsing.  The alternative would be to record the current
subject and predicate in the Turtle parse context, but this was more
trouble than I had time to bother with at the moment.